% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_daily.R
\name{summarise_daily}
\alias{summarise_daily}
\title{Summarise (sum or average) time series data at a daily scale}
\usage{
summarise_daily(.data, .timevar, .cumul, .average)
}
\arguments{
\item{.data}{a dataframe with the data to summarise}

\item{.timevar}{unquoted name of the column with date-time data}

\item{.cumul}{vector of unquoted name of columns with data to sum along each days}

\item{.average}{vector of unquoted name of columns with data to average along each days}
}
\value{
a dataframe with the requested columns summarised at a daily scale
}
\description{
Summarise (sum or average) time series data at a daily scale
}
\examples{
#find an example
}
